requirements.txt
numpy
scipy
opencv-python-headless
shapely
geopandas

matplotlib
pydantic

uvicorn
fastapi
azure-storage-blob